+++ tests/layer/test_knowledge_graph.py
@@ -24,7 +24,7 @@ import numpy as np
 from tensorflow.keras import Model, initializers, losses
 
 from stellargraph.mapper.knowledge_graph import KGTripleGenerator
-from stellargraph.layer.knowledge_graph import ComplEx
+from stellargraph.layer.knowledge_graph import ComplEx, DistMult
 
 from .. import test_utils
 from ..test_utils.graphs import knowledge_graph
@@ -79,6 +79,44 @@ def test_complex(knowledge_graph):
     w_r = edge_types[r_idx, :]
     e_o = nodes[o_idx, :]
     actual = (e_s * w_r * e_o.conj()).sum(axis=1).real
+
+    # predict every edge using the model
+    prediction = model.predict(gen.flow(df))
+
+    # (use an absolute tolerance to allow for catastrophic cancellation around very small values)
+    assert np.allclose(prediction[:, 0], actual, rtol=1e-3, atol=1e-14)
+
+
+def test_dismult(knowledge_graph):
+    # this test creates a random untrained model and predicts every possible edge in the graph, and
+    # compares that to a direct implementation of the scoring method in the paper
+    gen = KGTripleGenerator(knowledge_graph, 3)
+
+    # use a random initializer with a large positive range, so that any differences are obvious
+    init = initializers.RandomUniform(-1, 1)
+    x_inp, x_out = DistMult(gen, 5, embedding_initializer=init).build()
+
+    model = Model(x_inp, x_out)
+
+    every_edge = itertools.product(
+        knowledge_graph.nodes(),
+        knowledge_graph._edges.types.pandas_index,
+        knowledge_graph.nodes(),
+    )
+    df = triple_df(*every_edge)
+
+    # compute the exact values based on the model by extracting the embeddings for each element and
+    # doing the y_(e_1)^T M_r y_(e_2) = <e_1, w_r, e_2> inner product
+    s_idx = knowledge_graph._get_index_for_nodes(df.source)
+    r_idx = knowledge_graph._edges.types.to_iloc(df.label)
+    o_idx = knowledge_graph._get_index_for_nodes(df.target)
+
+    nodes, edge_types = DistMult.embeddings(model)
+    # the rows correspond to the embeddings for the given edge, so we can do bulk operations
+    e_s = nodes[s_idx, :]
+    w_r = edge_types[r_idx, :]
+    e_o = nodes[o_idx, :]
+    actual = (e_s * w_r * e_o).sum(axis=1)
 
     # predict every edge using the model
     prediction = model.predict(gen.flow(df))